Create README.md
Random Guesser game documentation
Project overview

The Random Guesser Game is a web-based application in which users guess a randomly generated number between 1 and 100. Players have a limited number of attempts to guess the correct number. And the application will reply if the guessed number is too high or too low.